+++ CHANGELOG.md
@@ -4,6 +4,9 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [Unreleased]
+## [0.5.6] - 2026-04-18
+- Added Physics tools for calculating gravity/jump force
+- Added warning to music manager if there's no music
 
 ## [0.5.5] - 2026-04-09
 - Change to List Extensions to use RandomAndNoise which uses System.Random instead of UnityEngine.Random to allow things like Shuffle/MutateSelf to work outside of things like Start/Awake